+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 2019.5.1
 ========
 
+**Features:**
+- Added #1171: Show start writing message when no posts.
+- Added #1173: Discord link to dashboard.
+
 **Bug Fixes:**
 - Fixed #1155: Fix top margin on main wrapper.
 - Fixed #1153: Fix dark mode on reader and post/page edition.